+++ CHANGELOG.md
@@ -24,7 +24,11 @@
     - The scss folder is now located at scss and no longer in src/scss
     - The package is now distributed as umd bundle and  esm bundle. The bundle file is bundle/angular2-mdl.js. The esm entry point is components/index.js
 * The module *MdlRootModule* is a replacement for *MdlModule*. The *MdlRootModule* makes sure that all providers are loaded into the root injector of the app. You can still use *MdlModule*. But you should be aware that the providers are possibly not singletons. You also can use *MdlModule.forRoot*. This is the same as *MdlRootModule*.
-
+* MdlDialogService and MdlSnackbarService have changed:
+    - there is no longer a setDefaultViewContainerRef method. There are two ways to specify a ViewContainerRef where dynamic Components will be attached:
+      - use MdlDialogOutletSevrice and setDefaultViewContainerRef
+      - use the <dialog-outlet></dialog-outlet> component. This component can be anywhere in your html site. the best place is the last child of the body element
+      - the possibility to ste a viewcontainerref for each dialog or snackbar has been removed
 <a name="1.8.2"></a>
 # 1.8.2 (2016-10-06)